Fix latent bugs tripped by -Wall and the fuller feature set
- tclexpat.c: declare `nread` as `ssize_t` so `nread < 0` actually
  catches `read()` errors (was `size_t`, so the check never fired and
  `(int)nread` with a wrapped SIZE_MAX was fed to XML_ParseBuffer).
- domjson.c: bump `char buf[20]` to `char buf[32]` — a 64-bit Tcl_Size
  formatted as `-2^63` is 20 chars before the NUL terminator.
- domalloc.c: include <dom.h> so `domLength` resolves.  Without it the
  extension fails to build with --enable-tdomalloc on Tcl 9.
- domalloc.c: cast through `uintptr_t` (not `unsigned int`) in the two
  hash-index computations.  The original 2002 cast silently dropped the
  upper 32 bits of the pointer on 64-bit systems, aliasing blocks that
  differed only in those bits onto the same hash bucket.  Preserves the
  32-bit behaviour and uses the full pointer on 64-bit.
- xmlsimple.c: two error-recovery paths freed a `domAlloc`-allocated
  attrnode via libc `FREE`, which under --enable-tdomalloc handed the
  middle of a block-allocator slab to glibc and corrupted the heap.
  Use `domFree` to match the allocator.
- tcldom.c: swap the Tcl 9-removed `Tcl_Backslash` for `Tcl_UtfBackslash`
  (tcldom_unknownCmd, only reached with --enable-unknown).
- aolstub.cpp: `NsTdom_Init`'s second arg changes from `void *` to
  `const void *` to match AOLserver 4.x `Ns_TclInterpInitProc`.

+++ generic/domalloc.c
@@ -38,6 +38,7 @@
 #include <tcl.h>
 #include <stdlib.h>
 #include <string.h>
+#include <dom.h>       /* for domLength typedef */
 #include <domalloc.h>
 
 
@@ -136,7 +137,7 @@ fillHashTable (
     domAllocBlock * hashedBlock;
     unsigned int    i;
 
-    i = ( (unsigned int)mem >> BLOCK_SIZE_BITS) & BIN_HASH_MASK;
+    i = ( (uintptr_t)mem >> BLOCK_SIZE_BITS) & BIN_HASH_MASK;
     hashedBlock = bins.hashedBlocks[i];
     while (hashedBlock != NULL) {
         if (hashedBlock == block) {
@@ -390,7 +391,7 @@ domFree (
     |
     \------------------------------------------------------------------*/
     if (!foundInCache) {
-        i = ( (unsigned int)mem >> BLOCK_SIZE_BITS) & BIN_HASH_MASK;
+        i = ( (uintptr_t)mem >> BLOCK_SIZE_BITS) & BIN_HASH_MASK;
         block = bins.hashedBlocks[i];
         while (block != NULL) {
             if ((mem > (void*)block) && (mem < (void*)(block->end))) break;
